refactor: refactored 3  api endpoints to seperate logic between controller and services
