+++ src/app/api/course-list/route.ts
@@ -1,14 +1,11 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/database/mongoose";
-import { Course } from "@/db/course";
+import { getCourseList } from "@/lib/services/subject";
 
 export const dynamic = "force-dynamic";
 
 export async function GET() {
   try {
-    await connectToDatabase();
-    const courses = await Course.find().lean();
-
+    const courses = await getCourseList();
     return NextResponse.json(courses, { status: 200 });
   } catch (error) {
     console.error(error);
